refactor(crypto): extract CryptoKeysLookup to cryptowiring (holomush-5iaov)
